pyproject.toml
[tool.poetry]
name = "danieldup"
version = "0.1.0"
description = ""
authors = ["DanielDuP <d.duplessis@sakeliga.co.za>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
matplotlib = "^3.9.1"
numpy = "^2.0.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
